bpf: tests: condense TC DSR-LB test
Run the IPv4 and IPv6 test from a single file.
